extract most all of steps to abstract xls importer class
